packages/pyllm/pyproject.toml
[project]
name = "llm-service"
version = "0.1.0"
description = "Flexible Docker-based LLM service supporting multiple models"
authors = [
  { name = "Camilla Thorsen", email = "camilla@thorsen.dev" }
]
license = { text = "UNLICENSED" }
requires-python = ">=3.11"

dependencies = [
  "fastapi (>=0.115.12,<0.116.0)",
  "uvicorn[standard] (>=0.34.3,<0.35.0)",
  "python-multipart",
  "pydantic (>=2.0.0,<3.0.0)",
  "llama-cpp-python (>=0.2.72,<0.3.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
